move tests for upload and process_and_upload to separate files
